+++ internal/gitaly/service/repository/testhelper_test.go
@@ -48,6 +48,7 @@ func testMain(m *testing.M) int {
 }
 
 func TestWithRubySidecar(t *testing.T) {
+	t.Parallel()
 	cfg := testcfg.Build(t)
 
 	testhelper.ConfigureGitalyHooksBin(t, cfg)
@@ -102,6 +103,7 @@ func setupRepositoryWithWorkingtreeServiceWithRuby(t testing.TB, cfg config.Cfg,
 
 func setupRepositoryServiceWithRuby(t testing.TB, cfg config.Cfg, rubySrv *rubyserver.Server, opts ...testserver.GitalyServerOpt) (config.Cfg, *gitalypb.Repository, string, gitalypb.RepositoryServiceClient) {
 	client, serverSocketPath := runRepositoryService(t, cfg, rubySrv, opts...)
+	testhelper.ConfigureGitalyGit2GoBin(t, cfg)
 	cfg.SocketPath = serverSocketPath
 
 	repo, repoPath, cleanup := gittest.CloneRepoAtStorage(t, cfg, cfg.Storages[0], t.Name())