+++ internal/metadata/featureflag/feature_flags.go
@@ -10,7 +10,7 @@ type FeatureFlag struct {
 // NOTE: if you add a new feature flag please add it to the `All` list defined below.
 var (
 	// GoUpdateRemoteMirror enables the Go implementation of UpdateRemoteMirror
-	GoUpdateRemoteMirror = FeatureFlag{Name: "go_update_remote_mirror", OnByDefault: false}
+	GoUpdateRemoteMirror = FeatureFlag{Name: "go_update_remote_mirror", OnByDefault: true}
 	// FetchInternalRemoteErrors makes FetchInternalRemote return actual errors instead of a boolean
 	FetchInternalRemoteErrors = FeatureFlag{Name: "fetch_internal_remote_errors", OnByDefault: false}
 	// TxConfig enables transactional voting for SetConfig and DeleteConfig RPCs.
@@ -22,6 +22,16 @@ var (
 	// LFSPointersPipeline enables the alternative pipeline implementation of LFS-pointer
 	// related RPCs.
 	LFSPointersPipeline = FeatureFlag{Name: "lfs_pointers_pipeline", OnByDefault: false}
+	// GoSetConfig enables git2go implementation of SetConfig.
+	GoSetConfig = FeatureFlag{Name: "go_set_config", OnByDefault: false}
+	// CreateRepositoryFromBundleAtomicFetch will add the `--atomic` flag to git-fetch(1) in
+	// order to reduce the number of transactional votes.
+	CreateRepositoryFromBundleAtomicFetch = FeatureFlag{Name: "create_repository_from_bundle_atomic_fetch", OnByDefault: false}
+	ResolveConflictsWithHooks             = FeatureFlag{Name: "resolve_conflicts_with_hooks", OnByDefault: false}
+	// ReplicateRepositoryDirectFetch will cause the ReplicateRepository RPC to perform fetches
+	// via a direct call instead of doing an RPC call to its own server. This fixes calls of
+	// `ReplicateRepository()` in case it's invoked via Praefect with transactions enabled.
+	ReplicateRepositoryDirectFetch = FeatureFlag{Name: "replicate_repository_direct_fetch", OnByDefault: false}
 )
 
 // All includes all feature flags.
@@ -32,4 +42,8 @@ var All = []FeatureFlag{
 	TxRemote,
 	GoFindLicense,
 	LFSPointersPipeline,
+	GoSetConfig,
+	CreateRepositoryFromBundleAtomicFetch,
+	ResolveConflictsWithHooks,
+	ReplicateRepositoryDirectFetch,
 }